applications/telesys-api/telesys/enums.py
from enum import Enum

import strawberry


@strawberry.enum
class Interval(Enum):
    MINUTE = "1m"
    FIVE_MINUTES = "5m"
    FIFTEEN_MINUTES = "15m"
    HOUR = "1h"
    FOUR_HOURS = "4h"
    DAY = "1d"
    WEEK = "1w"
